анимация призов в отдельный класс
